SelfieListFragment has been created but need to implemented.
