+++ app/src/main/res/layout/activity_main.xml
@@ -14,13 +14,12 @@
         android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
         app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
 
-    <ListView
+    <fragment
+        android:id="@+id/fragment"
+        android:name="com.example.mehmet.dailyselfie.SelfieListFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:id="@+id/listView"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true" />
+        android:layout_gravity="center_horizontal" />
 
 
 </LinearLayout>
