+++ app/src/main/res/layout/selfie_item.xml
